Checker: split isFormat API
The isFormat method has been renamed to matchesSuffix and matchesFormat.
Instead of using a boolean flag (or SCIFIOConfig) to differentiate the
use of these methods, they simply are intended to be called based on the
needs of the consuming code.

matchesSuffix will never open the underlying dataset, while
matchesFormat will always read the source if suffix checking is not
sufficient.

Refactored the SCIFIOConfig checker configuration method to be more
meaningfully named, and is now used appropriately to decid which of
these methods to call.